Don't silently clip the range argument to int32 before storing it internally as an int64 (improve future compatibility with sampling from huge tensors.  Note that this CL isn't enough to permit huge tensors, because we also clip their size at other points in the process, but it's a step forward.) Change: 117953355

+++ tensorflow/core/kernels/range_sampler.h
@@ -36,7 +36,7 @@ class Env;
 // [0, range)
 class RangeSampler {
  public:
-  explicit RangeSampler(int range) : range_(range) { CHECK_GT(range_, 0); }
+  explicit RangeSampler(int64 range) : range_(range) { CHECK_GT(range_, 0); }
   virtual ~RangeSampler();
 
   // Sample a single value
